+++ test/IntegrationTests/OpAmpTests.cs
@@ -28,7 +28,7 @@ public class OpAmpTests : TestHelper
         SetEnvironmentVariable("OTEL_DOTNET_AUTO_OPAMP_SERVER_URL", $"http://localhost:{server.Port}/v1/opamp");
         SetEnvironmentVariable(
             "OTEL_RESOURCE_ATTRIBUTES",
-            "opamp.test=true,service.namespace=my-namespace,service.namespace.name=not-a-semantic-convention");
+            "opamp.test=true,service.namespace=my-namespace");
 
         AgentDescription? agentDescriptionFrame = null;
 
@@ -51,9 +51,6 @@ public class OpAmpTests : TestHelper
             agentDescriptionFrame.IdentifyingAttributes,
             a => a.Key == "service.namespace" && a.Value.StringValue == "my-namespace");
         Assert.Contains(agentDescriptionFrame.NonIdentifyingAttributes, a => a.Key == "opamp.test");
-        Assert.Contains(
-            agentDescriptionFrame.NonIdentifyingAttributes,
-            a => a.Key == "service.namespace.name" && a.Value.StringValue == "not-a-semantic-convention");
     }
 
     [Fact]