+++ CHANGELOG.md
@@ -19,8 +19,8 @@ This component adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.h
 
 ### Fixed
 
-- Use the standard `service.namespace` resource attribute when identifying an
-  application through OpAMP.
+- Use the standard `service.namespace` resource attribute instead of `service.namespace.name`
+  when identifying an application through OpAMP.
 
 ## [v1.16.0](https://github.com/open-telemetry/opentelemetry-dotnet-instrumentation/releases/tag/v1.16.0)